+++ src/definitions.ts
@@ -275,6 +275,7 @@ export interface ScanResult {
 export interface BluetoothLePlugin {
   initialize(options?: InitializeOptions): Promise<void>;
   isEnabled(): Promise<BooleanResult>;
+  requestEnable(): Promise<void>;
   enable(): Promise<void>;
   disable(): Promise<void>;
   startEnabledNotifications(): Promise<void>;
@@ -293,7 +294,7 @@ export interface BluetoothLePlugin {
   addListener(eventName: string, listenerFunc: (event: ReadResult) => void): PluginListenerHandle;
   addListener(eventName: 'onScanResult', listenerFunc: (result: ScanResultInternal) => void): PluginListenerHandle;
   connect(options: DeviceIdOptions & TimeoutOptions): Promise<void>;
-  createBond(options: DeviceIdOptions): Promise<void>;
+  createBond(options: DeviceIdOptions & TimeoutOptions): Promise<void>;
   isBonded(options: DeviceIdOptions): Promise<BooleanResult>;
   disconnect(options: DeviceIdOptions): Promise<void>;
   getServices(options: DeviceIdOptions): Promise<BleServices>;